rename RandomMutation -> SimpleMutation
