+++ src/main/java/jp/kusumotolab/kgenprog/CUILauncher.java
@@ -11,7 +11,7 @@ import jp.kusumotolab.kgenprog.ga.crossover.Crossover;
 import jp.kusumotolab.kgenprog.ga.crossover.FirstVariantSelectionStrategy;
 import jp.kusumotolab.kgenprog.ga.crossover.SecondVariantSelectionStrategy;
 import jp.kusumotolab.kgenprog.ga.mutation.Mutation;
-import jp.kusumotolab.kgenprog.ga.mutation.RandomMutation;
+import jp.kusumotolab.kgenprog.ga.mutation.SimpleMutation;
 import jp.kusumotolab.kgenprog.ga.mutation.selection.RouletteStatementSelection;
 import jp.kusumotolab.kgenprog.ga.mutation.selection.StatementSelection;
 import jp.kusumotolab.kgenprog.ga.selection.DefaultVariantSelection;
@@ -42,7 +42,7 @@ public class CUILauncher {
     final Random random = new Random(config.getRandomSeed());
     final StatementSelection rouletteStatementSelection =
         new RouletteStatementSelection(random);
-    final Mutation mutation = new RandomMutation(config.getMutationGeneratingCount(), random,
+    final Mutation mutation = new SimpleMutation(config.getMutationGeneratingCount(), random,
         rouletteStatementSelection, config.getScope(), config.getNeedHistoricalElement());
     final FirstVariantSelectionStrategy firstVariantSelectionStrategy =
         config.getFirstVariantSelectionStrategy()
