+++ src/test/java/jp/kusumotolab/kgenprog/KGenProgMainTest.java
@@ -16,7 +16,7 @@ import jp.kusumotolab.kgenprog.ga.crossover.FirstVariantRandomSelection;
 import jp.kusumotolab.kgenprog.ga.crossover.SecondVariantRandomSelection;
 import jp.kusumotolab.kgenprog.ga.crossover.SinglePointCrossover;
 import jp.kusumotolab.kgenprog.ga.mutation.Mutation;
-import jp.kusumotolab.kgenprog.ga.mutation.RandomMutation;
+import jp.kusumotolab.kgenprog.ga.mutation.SimpleMutation;
 import jp.kusumotolab.kgenprog.ga.mutation.selection.CandidateSelection;
 import jp.kusumotolab.kgenprog.ga.mutation.selection.RouletteStatementSelection;
 import jp.kusumotolab.kgenprog.ga.selection.GenerationalVariantSelection;
@@ -53,7 +53,7 @@ public class KGenProgMainTest {
     final FaultLocalization faultLocalization = new Ochiai();
     final Random random = new Random(config.getRandomSeed());
     final CandidateSelection statementSelection = new RouletteStatementSelection(random);
-    final Mutation mutation = new RandomMutation(config.getMutationGeneratingCount(), random,
+    final Mutation mutation = new SimpleMutation(config.getMutationGeneratingCount(), random,
         statementSelection, config.getScope(), true);
     final Crossover crossover =
         new SinglePointCrossover(random, new FirstVariantRandomSelection(random),
